+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * improved css icons
 * changed assets version to time in debug mode and moved version info to `package.json`
 * replaced `.htaccess` with `.htaccess.dist`
+* introduced theme config file
+* improved Lexy settings
 
 
 ## 0.1.4